add column refactored
add column refactored
