+++ src/components/FormsUI/TextField/index.js
@@ -1,21 +1,1 @@
-import React from 'react';
-import TextField from '@mui/material/TextField';
-import { useField } from 'formik';
-
-const TextFieldWrapper = ({ name, ...otherProps }) => {
-  const [field, meta] = useField(name);
-
-  const configTextfield = {
-    ...field,
-    ...otherProps,
-    fullWidth: true,
-    variant: 'outlined',
-  };
-  if (meta && meta.touched && meta.error) {
-    configTextfield.error = true;
-    configTextfield.helperText = meta.error;
-  }
-
-  return <TextField {...configTextfield} />;
-};
-export default TextFieldWrapper;
+export { default } from './TextField';
